+++ frontend/src/components/forms/Autocomplete/index.jsx
@@ -2,7 +2,7 @@ import React, { memo, useCallback, useMemo, useState } from 'react';
 import { Select, Typography } from 'antd';
 import PropTypes from 'prop-types';
 import { useDebounce } from 'react-use';
-import styles from './Autocomplete.module.less';
+import styles from './styles.module.less';
 
 const { Text } = Typography;