+++ frontend/src/components/forms/NonFieldErrors/index.jsx
@@ -1,6 +1,6 @@
 import React, { memo } from 'react';
 import { Form as AntdForm } from 'antd';
-import styles from './NonFieldErrors.module.less';
+import styles from './styles.module.less';
 
 const NonFieldErrors = () => (
   <AntdForm.Item className={styles.nonFieldErrors} name="nonFieldErrors" />