+++ frontend/src/components/forms/components/index.jsx
@@ -1,2 +1,0 @@
-export { default as FormCard } from './FormCard';
-export { default as NonFieldErrors } from './NonFieldErrors';