+++ frontend/src/components/ui/IconButton/index.jsx
@@ -2,7 +2,7 @@ import React, { memo } from 'react';
 import { Button, Tooltip } from 'antd';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import styles from './IconButton.module.less';
+import styles from './styles.module.less';
 
 const IconButton = ({
   icon,