+++ frontend/src/components/ui/Spin/index.jsx
@@ -1,7 +1,7 @@
 import React, { memo } from 'react';
 import { Spin as AntdSpin } from 'antd';
 import PropTypes from 'prop-types';
-import styles from './Spin.module.less';
+import styles from './styles.module.less';
 
 const Spin = ({ text = 'Loading...', children, spinning = true }) => (
   <div className={styles.spinContainer}>