+++ frontend/src/pages/LoginPage/index.jsx
@@ -1,10 +1,10 @@
 import React, { memo, useCallback, useState } from 'react';
 import { Col, Image, Row, Typography } from 'antd';
 import { jkdevLogoUrl } from '@/assets';
-import { LoginForm } from '@/components/forms';
+import { LoginForm } from '@/components/features/User';
 import { Spin } from '@/components/ui';
 import { useAuthStore } from '@/stores';
-import styles from './LoginPage.module.less';
+import styles from './styles.module.less';
 
 const { Text } = Typography;
 
@@ -39,9 +39,8 @@ const LoginPage = () => {
       <Col span={24}>
         <Text italic>
           {/* eslint-disable-next-line react/no-unescaped-entities */}
-          Bienvenue sur le site de Django React Starter. Veuillez-vous connecter
-          pour continuer. Pour créer un compte, merci de contacter un
-          administrateur.
+          Bienvenue sur le site de Hubby. Veuillez-vous connecter pour
+          continuer. Pour créer un compte, merci de contacter un administrateur.
         </Text>
       </Col>
       <Col span={24}>