+++ CHANGELOG.md
@@ -12,7 +12,7 @@
   - Reworked imports for components and pages
   - Updated theme colors
   - Update design for page banner and other small components
-  - Added some new components (`Autocomplete`, `GhostTag`, `Section`, `Switch`)
+  - Added some new components (`Autocomplete`, `GhostTag`, `Section`, `Switch`, `InputNumber`)
 - **Backend**
   - Fixed catch-all route in `urls.py`
   - Fixed Dockerfile not copying correctly the frontend build